# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "description": "Cette API permet de gérer les différentes composantes de la course d'orientation fixe de Bex.",
    "version": "1.0.0",
    "title": "API Course orientation Bex"
  },
  "servers": [
    {
      "url": "https://cobex-uu6t.onrender.com/api",
      "description": "URL online, elle peut mettre jusqu'à une minute pour répondre"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "URL local pour faire des tests"
    }
  ],
...
  "paths": {
...
    "/postes/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          },
          "description": "ID du poste à récupérer"
        }
      ],
      "get": {
        "summary": "Récupérer un poste par son ID",
        "tags": [
          "postes"
        ],
        "responses": {
          "200": {
            "description": "Poste récupéré avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartialPosteData"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/NotID"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "patch": {
        "summary": "Modifier l'accès à un poste (uniquement accessible aux superAdmin)",
        "tags": [
          "postes"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "description": "Indique si le poste est accessible actuellement",
              "schema": {
                "type": "object",
                "properties": {
                  "estAccessible": {
                    "type": "boolean",
                    "description": "Indique si le poste est accessible actuellement",
                    "example": false
                  }
                },
                "required": [
                  "estAccessible"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Poste modifié avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Poste"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "description": "Le code du QR du poste",
                          "example": "37ahj"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Aucune donnée à modifier",
            "content": {
              "text/html": {
                "example": "No data to update"
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/NotConnected"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/parcours": {
      "post": {
        "summary": "Créer un nouveau parcours (uniquement accessible aux rôles admin et superAdmin)",
        "tags": [
          "parcours"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrailData"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Parcours créé avec succès",
            "headers": {
              "Location": {
                "description": "URL du nouveau parcours",
                "schema": {
                  "type": "string",
                  "example": "http://localhost:4000;/api/parcours/00f8e6fedc4cf8b5ace5a689"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrailCreateResponse"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/NotConnected"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "description": "Parcours invalide",
            "content": {
              "text/html": {
                "example": "Trail validation failed : nom: Parcours 1 already exists"
              }
            }
          }
        }
      },
      "get": {
        "summary": "Récupérer la liste des parcours",
        "tags": [
          "parcours"
        ],
        "parameters": [
          {
            "name": "include",
            "in": "query",
            "description": "Inclure des informations supplémentaires (nom du créateur)",
            "schema": {
              "type": "string",
              "enum": [
                "user"
              ],
              "example": "user"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des parcours récupérée avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PartialTrailData"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/parcours/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          },
          "description": "ID de l'utilisateur à récupérer"
        }
      ],
      "get": {
        "summary": "Récupérer un parcours par son ID",
        "tags": [
          "parcours"
        ],
        "parameters": [
          {
            "name": "include",
            "in": "query",
            "description": "Inclure 4 résultats et les informations sur les postes",
            "schema": {
              "type": "string",
              "enum": [
                "postes"
              ],
              "example": "postes"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Numéro de page (4 postes par page)",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des utilisateurs récupérée avec succès",
            "content": {
              "application/json": {
                "examples": {
                  "withoutInclude": {
                    "summary": "Sans paramètre de requête",
                    "value": {
                      "id": "678504c0224b7e662ab961ff",
                      "nom": "Parcours Facile"
                    }
                  },
                  "withInclude": {
                    "summary": "Avec paramètre `include=postes&page=3`",
                    "value": {
                      "id": "678504c0224b7e662ab961ff",
                      "nom": "Parcours Facile",
                      "difficulte": "facile",
                      "descr": "Un parcours facile pour les débutants.",
                      "postesInclus": [
                        {
                          "geoloc": {
                            "lat": 46.780534,
                            "long": 6.642302,
                            "_id": "678504c0224b7e662ab961f5",
                            "id": "678504c0224b7e662ab961f5"
                          },
                          "number": 30,
                          "estAccessible": true,
                          "images": [
                            "https://commons.wikimedia.org/wiki/File:For%C3%AAt_de_Tusson,_15.jpg"
                          ],
                          "descr": "Poste 1 (au coin de la Banque)",
                          "code": "30bda",
                          "createdAt": "2025-01-13T12:19:12.605Z",
                          "id": "678504c0224b7e662ab961f4"
                        },
                        {
                          "geoloc": {
                            "lat": 46.780906,
                            "long": 6.643686,
                            "_id": "678504c0224b7e662ab961f7",
                            "id": "678504c0224b7e662ab961f7"
                          },
                          "number": 31,
                          "estAccessible": true,
                          "images": [
                            "https://commons.wikimedia.org/wiki/File:For%C3%AAt_de_Tusson,_22.jpg",
                            "https://commons.wikimedia.org/wiki/File:For%C3%AAt_de_Tusson,_04.jpg"
                          ],
                          "descr": "Poste 2 (sur le pont)",
                          "code": "31ert",
                          "createdAt": "2025-01-13T12:19:12.605Z",
                          "id": "678504c0224b7e662ab961f6"
                        }
                      ],
                      "resultatsAct": [
                        {
                          "_id": "678504c0224b7e662ab9620f",
                          "temps": 190,
                          "user": "Jane Doe"
                        },
                        {
                          "_id": "678504c0224b7e662ab9620c",
                          "temps": 200,
                          "user": "Mary Doe"
                        },
                        {
                          "_id": "678504c0224b7e662ab96213",
                          "temps": 200,
                          "user": "Jane Doe"
                        },
                        {
                          "_id": "678504c0224b7e662ab96212",
                          "temps": 210,
                          "user": "Mary Doe"
                        }
                      ],
                      "nombrePages": 3
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/NotID"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "summary": "Modifier un parcours complet (accessible uniquement au créateur du parcours (admin) et aux superAdmin)",
        "tags": [
          "parcours"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrailData"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Parcours modifié avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Trail"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/NotID"
          },
          "401": {
            "$ref": "#/components/responses/NotConnected"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "description": "Parcours invalide",
            "content": {
              "text/html": {
                "example": "Trail validation failed : nom: Printemps 1 already exists"
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Modifier partiellement un parcours (accessible uniquement au créateur du parcours (admin) et aux superAdmin)",
        "tags": [
          "parcours"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "nom": {
                    "type": "string",
                    "description": "Nom du parcours",
                    "example": "Parcours 1"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Parcours modifié avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Trail"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/NotID"
          },
          "401": {
            "$ref": "#/components/responses/NotConnected"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "delete": {
        "summary": "Supprimer un parcours et les résultats associés (accessible uniquement au créateur du parcours (admin) et aux superAdmin)",
        "tags": [
          "parcours"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Parcours supprimé avec succès"
          },
          "400": {
            "$ref": "#/components/responses/NotID"
          },
          "401": {
            "$ref": "#/components/responses/NotConnected"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "responses": {
...
      "NotConnected": {
        "description": "Personne non connectée",
        "content": {
          "text/html": {
            "example": "Authorization header is missing"
          }
        }
      },
      "Forbidden": {
        "description": "Interdit (personne connectée mais pas avec le bon rôle)",
        "content": {
          "text/html": {
            "example": "Forbidden"
          }
        }
      },
      "NotFound": {
        "description": "Non trouvé",
        "content": {
          "text/html": {
            "example": "<STH> not found"
          }
        }
      },
...
      "NotID": {
        "description": "ID invalide",
        "content": {
          "text/html": {
            "example": "Invalid ID"
          }
        }
      }
    },
    "schemas": {
      "BasePerson": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Identifiant unique de l'utilisateur",
            "example": "7038e6fedc4cf8b5ace5a689"
          },
          "nom": {
            "type": "string",
            "description": "Nom de l'utilisateur",
            "minLength": 3,
            "maxLength": 30,
            "example": "John"
          }
        }
      },
      "SavedPerson": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BasePerson"
          },
          {
            "type": "object",
            "properties": {
              "mail": {
                "type": "mail",
                "description": "Mail de l'utilisateur",
                "example": "example@example.com",
                "format": "email"
              },
              "role": {
                "type": "string",
                "description": "Rôle de l'utilisateur",
                "enum": [
                  "utilisateur",
                  "admin",
                  "superAdmin"
                ],
                "example": "utilisateur"
              },
              "createdAt": {
                "type": "string",
                "description": "Date de création de l'utilisateur",
                "format": "date-time",
                "example": "2022-01-01T00:00:00Z"
              }
            }
          }
        ]
      },
      "Person": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SavedPerson"
          },
          {
            "type": "object",
            "properties": {
              "mdp": {
                "type": "string",
                "description": "Mot de passe de l'utilisateur",
                "example": "$2b$10$IAIz7tCzFRbES144XpFnWuloD5i/1crzvntErVJZ1AbGRPQNyqQym",
                "format": "password"
              }
            }
          }
        ]
      },
...
      "Geoloc": {
        "type": "object",
        "properties": {
          "lat": {
            "type": "integer",
            "description": "Latitude du poste",
            "example": 46.781278
          },
          "long": {
            "type": "integer",
            "description": "Longitude du poste",
            "example": 6.646287
          },
          "id": {
            "type": "string",
            "description": "Identifiant unique de la géolocalisation",
            "example": "d038e6fedc4cf8b5ace5a123"
          },
          "_id": {
            "type": "string",
            "description": "Identifiant unique de la géolocalisation",
            "example": "d038e6fedc4cf8b5ace5a123"
          }
        },
        "description": "Coordonnées du poste"
      },
      "BasePoste": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Identifiant unique du poste",
            "example": "d038e6fedc4cf8b5ace5a689"
          },
          "number": {
            "type": "integer",
            "description": "Le numéro de poste tel que inscrit sur la carte globale",
            "minimum": 30,
            "exclusiveMinimum": true,
            "example": 37
          },
          "geoloc": {
            "$ref": "#/components/schemas/Geoloc"
          }
        },
        "required": [
          "id",
          "geoloc",
          "number"
        ]
      },
      "Poste": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BasePoste"
          },
          {
            "type": "object",
            "properties": {
              "estAccessible": {
                "type": "boolean",
                "description": "Indique si le poste est accessible actuellement",
                "example": true
              },
              "images": {
                "type": "array",
                "description": "Toutes les images qui sont liées au poste",
                "items": {
                  "type": "string",
                  "format": "url"
                }
              },
              "descr": {
                "type": "string",
                "description": "Description du poste",
                "example": "A gauche de ce poste se trouve l'arbre centenaire..."
              },
              "createdAt": {
                "type": "string",
                "description": "Date de création du poste",
                "format": "date-time",
                "example": "2022-01-01T00:00:00Z"
              }
            },
            "required": [
              "estAccessible",
              "createdAt"
            ]
          }
        ]
      },
      "PartialPosteData": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BasePoste"
          },
          {
            "type": "object",
            "properties": {
              "estAccessible": {
                "type": "boolean",
                "description": "Indique si le poste est accessible actuellement",
                "example": true
              },
              "images": {
                "type": "array",
                "description": "Toutes les images qui sont liées au poste",
                "items": {
                  "type": "string",
                  "format": "url"
                }
              },
              "descr": {
                "type": "string",
                "description": "Description du poste",
                "example": "A gauche de ce poste se trouve l'arbre centenaire..."
              },
              "code": {
                "type": "string",
                "description": "Le code du QR du poste",
                "example": "37ahj"
              }
            }
          }
        ]
      },
...
      "BaseTrail": {
        "type": "object",
        "properties": {
          "nom": {
            "type": "string",
            "description": "Nom du parcours",
            "minLength": 3,
            "maxLength": 30,
            "example": "Printemps 1"
          },
          "difficulte": {
            "type": "string",
            "description": "Difficulté du parcours",
            "enum": [
              "facile",
              "moyen",
              "difficle",
              "très difficle"
            ],
            "example": "facile"
          },
          "descr": {
            "type": "string",
            "description": "Descriptif du parcours",
            "example": "Ce parcours peut se faire en famille avec une poucette..."
          }
        },
        "required": [
          "nom",
          "difficulte"
        ]
      },
      "Trail": {
        "allOf": [
          {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "description": "Identifiant unique du parcours",
                "example": "00f8e6fedc4cf8b5ace5a689"
              }
            }
          },
          {
            "$ref": "#/components/schemas/BaseTrail"
          },
          {
            "type": "object",
            "properties": {
              "createdAt": {
                "type": "string",
                "description": "Date de création du parcours",
                "format": "date-time",
                "example": "2022-01-01T00:00:00Z"
              },
              "createdBy": {
                "type": "string",
                "description": "L'identifiant unique de la personne qui a créé le parcours",
                "example": "7038e6fedc4cf8b5ace5a689"
              },
              "postesInclus": {
                "type": "array",
                "description": "La liste des postes qui sont dans le parcours",
                "items": {
                  "type": "string"
                },
                "example": [
                  "d038e6fedc4cf8b5ace5a689",
                  "d027e6fedc4cf8b5ace5a646"
                ]
              }
            }
          }
        ]
      },
      "TrailData": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseTrail"
          },
          {
            "type": "object",
            "properties": {
              "postesInclus": {
                "type": "array",
                "description": "La liste des postes qui sont dans le parcours",
                "items": {
                  "type": "string"
                },
                "example": [
                  "d038e6fedc4cf8b5ace5a689",
                  "d027e6fedc4cf8b5ace5a646"
                ]
              }
            },
            "required": [
              "postesInclus"
            ]
          }
        ]
      },
      "PartialTrailData": {
        "allOf": [
          {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "description": "Identifiant unique du parcours",
                "example": "00f8e6fedc4cf8b5ace5a689"
              }
            }
          },
          {
            "$ref": "#/components/schemas/BaseTrail"
          },
          {
            "type": "object",
            "properties": {
              "nbr_posts": {
                "type": "integer",
                "description": "Le nombre des postes qui sont dans le parcours",
                "example": 2
              }
            }
          }
        ]
      },
      "TrailCreateResponse": {
        "type": "object",
        "allOf": [
          {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "description": "Identifiant unique du parcours",
                "example": "00f8e6fedc4cf8b5ace5a689"
              }
            }
          },
          {
            "$ref": "#/components/schemas/TrailData"
          },
          {
            "type": "object",
            "properties": {
              "createdAt": {
                "type": "string",
                "description": "Date de création du parcours",
                "format": "date-time",
                "example": "2022-01-01T00:00:00Z"
              },
              "createdBy": {
                "type": "object",
                "allOf": [
                  {
                    "$ref": "#/components/schemas/Person"
                  }
                ]
              }
            }
          }
        ]
      }
    }
  }
}